pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragicamp"
version = "0.1.0"
description = "A modular framework for experimenting with RAG approaches"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "RAGiCamp Team"}
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.10.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
metrics = [
    "bert-score>=0.3.13",
    "bleurt @ git+https://github.com/google-research/bleurt.git",
]
viz = [
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "wandb>=0.15.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

